Add files via upload
These mini programs are created for "The Authentic Burger and Bakery Restaurant" to engage customers. The programs provided include a greeting to the customer, allowing them to choose a seat, review the menu(s) and order a from the Main Menu or the Build-your-Own-Burger Menu. After ordering, the customers receive a receipt and are encouraged to check out our customer Review page while leaving a comment! Lastly, the opening times of the Restaurant are provided to inform customers of our daily operation times.